:recycle: Refactor: 리뷰 반영 (1)
- 폰트 파일 이동
- modalLogo, IconButton 삭제 [YW-144]
