+++ src/components/common/Button/index.ts
@@ -1,2 +1,1 @@
 export * from "./Button";
-export * from "./IconButton";
