+++ src/components/common/Icon/assets.ts
@@ -13,7 +13,6 @@ export { default as logo } from "/public/icon/logo.svg";
 export { default as meatball } from "/public/icon/meatball.svg";
 export { default as memeShare } from "/public/icon/memeShare.svg";
 export { default as menu } from "/public/icon/menu.svg";
-export { default as modalLogo } from "/public/icon/modalLogo.svg";
 export { default as loginprofile } from "/public/icon/profile.svg";
 export { default as notloginprofile } from "/public/icon/profile2.svg";
 export { default as search } from "/public/icon/search.svg";
