+++ src/components/meme/MemeInfo/Button/index.ts
@@ -1,4 +1,3 @@
 export * from "./ClipboardCopyButton";
-export * from "./DownloadButton";
 export * from "./KakaoShareButton";
 export * from "./NativeShareButton";
