+++ src/styles/fonts/index.ts
@@ -1,0 +1,18 @@
+import localFont from "@next/font/local";
+
+export const pretendard = localFont({
+  src: "./PretendardVariable.woff2",
+  variable: "--font-pretendardVariable",
+  weight: "100 900",
+});
+
+export const suit = localFont({
+  src: "./SUIT-Variable.woff2",
+  variable: "--font-suitVariable",
+  weight: "100 900",
+});
+
+export const tossface = localFont({
+  src: "./TossFaceFontMac.ttf",
+  variable: "--font-tossfaceVariable",
+});
